+++ app/contexts/PromobarProvider/usePromobarContext.ts
@@ -1,0 +1,10 @@
+import {createContext, useContext} from 'react';
+
+import type {PromobarContext} from '~/lib/types';
+
+export const Context = createContext({
+  state: {},
+  actions: {},
+} as PromobarContext);
+
+export const usePromobarContext = () => useContext(Context);
