+++ app/contexts/SettingsProvider/useSettingsContext.ts
@@ -1,0 +1,10 @@
+import {createContext, useContext} from 'react';
+
+import type {SettingsContext} from '~/lib/types';
+
+export const Context = createContext({
+  state: {},
+  actions: {},
+} as SettingsContext);
+
+export const useSettingsContext = () => useContext(Context);
