+++ app/contexts/index.ts
@@ -1,7 +1,1 @@
-export * from './AnalyticsProvider';
-export * from './ContextsProvider';
-export * from './GlobalProvider';
-export * from './GroupingsProvider';
-export * from './MenuProvider';
-export * from './PromobarProvider';
-export * from './SettingsProvider';
+export {ContextsProvider} from './ContextsProvider';
